Moved hermit commands code to separate file
